pyproject.toml
[project]
name = "langgraph-playground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-community>=0.3.24",
    "langchain-core>=0.3.60",
    "langchain-openai>=0.3.17",
    "langchain-tavily>=0.1.6",
    "langgraph>=0.4.5",
    "langgraph-cli[inmem]>=0.2.10",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
]
